third_party/blink/web_tests/serial/serialPort_readable.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/mojo/public/mojom/base/unguessable_token.mojom.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/serial/serial.mojom.js"></script>
<script src="resources/serial-test-utils.js"></script>
<script>

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);

  assert_equals(port.readable, null);

  await port.open({ baudrate: 9600 });
  const readable = port.readable;
  assert_true(readable instanceof ReadableStream);

  port.close();
  assert_equals(port.readable, null);

  const reader = readable.getReader();
  const { value, done } = await reader.read();
  assert_true(done);
  assert_equals(value, undefined);
}, 'SerialPort.readable is set by open() and closes on port close');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  // Select a buffer size larger than the amount of data transferred.
  await port.open({ baudrate: 9600, buffersize: 64 });

  const data = new Uint8Array([1, 2, 3, 4, 5, 6, 7, 8]);
  fakePort.write(data);

  const reader = port.readable.getReader();
  let { value, done } = await reader.read();
  assert_false(done);
  compareArrays(data, value);

  port.close();

  ({ value, done } = await reader.read());
  assert_true(done);
  assert_equals(undefined, value);
}, 'Can read a small amount of data');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  // Select a buffer size smaller than the amount of data transferred.
  await port.open({ baudrate: 9600, buffersize: 64 });

  const data = new Uint8Array(1024);  // Much larger than buffersize above.
  for (let i = 0; i < data.byteLength; ++i)
    data[i] = i & 0xff;
  fakePort.write(data);

  const reader = port.readable.getReader();
  port.close();
  const value = await readAll(reader);
  compareArrays(data, value);
}, 'Can read a large amount of data');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  // Select a buffer size smaller than the amount of data transferred.
  await port.open({ baudrate: 9600, buffersize: 64 });

  const data = new Uint8Array([1, 2, 3, 4, 5, 6, 7, 8]);
  fakePort.write(data);
  fakePort.simulateParityError();

  let readable = port.readable;
  let reader = readable.getReader();
  let { value, done } = await reader.read();
  assert_false(done);
  compareArrays(data, value);

  await promise_rejects(t, new DOMException('', 'ParityError'), reader.read());
  assert_not_equals(port.readable, readable);

  readable = port.readable;
  assert_true(readable instanceof ReadableStream);
  reader = port.readable.getReader();

  await fakePort.waitForErrorCleared();
  fakePort.write(data);

  ({ value, done } = await reader.read());
  assert_false(done);
  compareArrays(data, value);

  port.close();
  assert_equals(port.readable, null);

  ({ value, done } = await reader.read());
  assert_true(done);
  assert_equals(undefined, value);
}, 'Parity error closes readable and replaces it with a new stream');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  // Select a buffer size larger than the amount of data transferred.
  await port.open({ baudrate: 9600, buffersize: 64 });

  fakePort.write(new TextEncoder().encode("Hello world!"));

  const reader = port.readable.pipeThrough(new TextDecoderStream()).getReader();
  let { value, done } = await reader.read();
  assert_false(done);
  assert_equals("Hello world!", value);

  port.close();

  ({ value, done } = await reader.read());
  assert_true(done);
  assert_equals(undefined, value);
}, 'Can pipe readable through a transform stream.')

</script>
